[Rename] #42 기존 extension 파일들을 util 패키지에서 extension 패키지로 이동
- DialogUtil 을 생성하며 기존에 util 패키지에 존재하던 extension 파일들과는 성질이 다르다고 판단하여 패키지 분리
