+++ presentation/src/main/java/com/woowahan/banchan/ui/maindishbanchan/MainDishBanchanFragment.kt
@@ -11,10 +11,10 @@ import com.woowahan.banchan.R
 import com.woowahan.banchan.databinding.FragmentMainDishBanchanBinding
 import com.woowahan.banchan.ui.base.BaseFragment
 import com.woowahan.banchan.ui.viewmodel.MainDishBanchanViewModel
-import com.woowahan.banchan.util.dp
-import com.woowahan.banchan.util.repeatOnStarted
-import com.woowahan.banchan.util.showSnackBar
-import com.woowahan.banchan.util.showToast
+import com.woowahan.banchan.extension.dp
+import com.woowahan.banchan.extension.repeatOnStarted
+import com.woowahan.banchan.extension.showSnackBar
+import com.woowahan.banchan.extension.showToast
 import com.woowahan.domain.model.BanchanModel
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
